+++ src/envs/carl_env.py
@@ -5,7 +5,7 @@ import numpy as np
 import os
 import json
 from typing import Dict, Tuple, Union, List, Optional, Any
-from src.context.context_changer import add_gaussian_noise
+from src.context.augmentation import add_gaussian_noise
 from src.context.context_utils import get_context_bounds
 from src.trial_logger import TrialLogger
